requirements.txt
discord.py==1.7.3
simpleBDB==1.1.28